+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Changelog
 =========
 
 - Add ``NestedSetting`` for easy management of nested settings.
+- Deprecate setting checkers in favor of validators, similarly to Django form fields.
 
 0.3.0 (2017-11-30)
 ==================